+++ ssl.t.c
@@ -29,13 +29,22 @@
 #include "base64.h"
 #include "ssl.h"
 
+#include <limits.h>
 #include <stdlib.h>
+#include <unistd.h>
 
 #include <check.h>
+
+#ifdef __APPLE__
+#define DLSUFFIX "dylib"
+#else
+#define DLSUFFIX "so"
+#endif
 
 #define TESTKEY "extra/pki/server.key"
 #define TESTCERT "extra/pki/server.crt"
 #define TESTCERT2 "extra/pki/rsa.crt"
+#define ENGINE "extra/engine/dummy-engine."DLSUFFIX
 
 static void
 ssl_setup(void)
@@ -748,6 +757,21 @@ START_TEST(ssl_x509_refcount_inc_01)
 }
 END_TEST
 
+#ifndef OPENSSL_NO_ENGINE
+START_TEST(ssl_engine_01)
+{
+	char cwd[PATH_MAX];
+	char *path;
+
+	fail_unless(getcwd(cwd, sizeof(cwd)) == cwd, "getcwd() failed");
+	fail_unless(asprintf(&path, "%s/"ENGINE, cwd) != -1 && !!path,
+	            "constructing engine path failed");
+	fail_unless(ssl_engine(path) == 0, "loading OpenSSL engine failed");
+	free(path);
+}
+END_TEST
+#endif /* !OPENSSL_NO_ENGINE */
+
 Suite *
 ssl_suite(void)
 {
@@ -850,6 +874,16 @@ ssl_suite(void)
 	tcase_add_checked_fixture(tc, ssl_setup, ssl_teardown);
 	tcase_add_test(tc, ssl_x509_refcount_inc_01);
 	suite_add_tcase(s, tc);
+
+#ifndef OPENSSL_NO_ENGINE
+	tc = tcase_create("ssl_engine");
+	tcase_add_checked_fixture(tc, ssl_setup, ssl_teardown);
+	tcase_add_test(tc, ssl_engine_01);
+	suite_add_tcase(s, tc);
+#else /* OPENSSL_NO_ENGINE */
+	fprintf(stderr, "1 test omitted because OpenSSL has no "
+	                "engine support\n");
+#endif /* OPENSSL_NO_ENGINE */
 
 	return s;
 }