+++ NEWS.md
@@ -3,6 +3,7 @@
 
 -   Suppress Expect-CT header in order to avoid Certificate Transparency log
     lookup failures (issue #205).
+-   Add -x option for activating an OpenSSL engine (issue #204, pull req #206).
 -   Add -f option for loading configuration from file (pull req #193).
 -   Add `sudotest` target with unit tests which require privileges to run.
 -   Add XNU headers for macOS High Sierra 10.13.1, 10.13.2 and 10.13.3.